// pages/auth/type.ts
export interface OwnProp {
  name: string | undefined;
  family: string | undefined;
  password: string | undefined;
  repassword: string | undefined;
  username: string | undefined;
  email: string | undefined;
  address: string | undefined;
  phone: string | undefined;
  sex: string | undefined;
}